[W9D2] JWT token 적용, 좋아요 API 수정

+++ controller/CartController.js
@@ -1,8 +1,7 @@
+const ensureAuthorization = require("../auth");
 const jwt = require("jsonwebtoken");
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
-const dotenv = require("dotenv");
-dotenv.config();
 
 const addToCart = (req, res) => {
   const { book_id, quantity } = req.body;
@@ -47,8 +46,16 @@ const getCartItems = (req, res) => {
     let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
     FROM cartItems LEFT JOIN books 
     ON cartItems.book_id = books.id
-    WHERE user_id = ? AND cartItems.id IN (?)`;
-    conn.query(sql, [authorization.id, selected], (err, results) => {
+    WHERE user_id = ? `;
+    let values = [authorization.id];
+
+    if (selected) {
+      // 선택한 장바구니 목록 조회
+      sql += `AND cartItems.id IN (?)`;
+      values.push(selected);
+    }
+
+    conn.query(sql, values, (err, results) => {
       if (err) {
         console.log(err);
         return res.status(StatusCodes.BAD_REQUEST).end();
@@ -59,33 +66,28 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-  const cartItemId = req.params.id;
+  let authorization = ensureAuthorization(req);
 
-  let sql = `DELETE FROM cartItems WHERE id = ?`;
-  conn.query(sql, cartItemId, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
-};
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    const cartItemId = req.params.id;
 
-function ensureAuthorization(req, res) {
-  try {
-    let receivedJwt = req.headers["authorization"];
-    console.log(receivedJwt);
-
-    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-    console.log(decodedJwt);
-
-    return decodedJwt;
-  } catch (err) {
-    console.log(err.name);
-    console.log(err.message);
-
-    return err;
+    let sql = `DELETE FROM cartItems WHERE id = ?`;
+    conn.query(sql, cartItemId, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
   }
-}
+};
 
 module.exports = { addToCart, getCartItems, removeCartItem };
